feat: planner and user delete

+++ backend/src/main/java/com/mysite/weddingyou_backend/plannerUpdateDelete/PlannerUpdateDeleteService.java
@@ -1,5 +1,6 @@
 package com.mysite.weddingyou_backend.plannerUpdateDelete;
 
+import java.util.List;
 import java.util.Optional;
 import java.util.Properties;
 import java.util.Random;
@@ -16,9 +17,12 @@ import javax.mail.internet.MimeMessage;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.mysite.weddingyou_backend.like.LikeEntity;
+import com.mysite.weddingyou_backend.like.LikeRepository;
 import com.mysite.weddingyou_backend.plannerRegister.PlannerRegister;
 import com.mysite.weddingyou_backend.userUpdateDelete.UserUpdateDelete;
 
+import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 
 
@@ -28,6 +32,9 @@ public class PlannerUpdateDeleteService {
 	 
 	@Autowired
 	private PlannerUpdateDeleteRepository plannerRepository;
+
+	@Autowired
+	private LikeRepository likeRepository;
 	
 	public PlannerUpdateDelete getPlannerByEmail(String email) {
 		//System.out.println(plannerRepository.findByEmail(email));
@@ -40,10 +47,15 @@ public class PlannerUpdateDeleteService {
 		// repository의 save 메서드 호출(조건. entity 객체를 넘겨줘야 함)
 	}
 	
-	public void delete(PlannerUpdateDelete planner) {
-		//repository의 delete 메소드 소환
-		plannerRepository.delete(planner);
-		// repository의 delete 메서드 호출(조건. entity 객체를 넘겨줘야 함)
+	public void delete(PlannerUpdateDelete planner) { 
+    if (planner != null) {
+        List<LikeEntity> likes = likeRepository.findByPlannerEmail(planner.getEmail());
+        likeRepository.deleteAll(likes);
+        
+        plannerRepository.delete(planner);
+    } else {
+        throw new EntityNotFoundException("Planner with email not found.");
+    }
 	}
